multi-label/requirements.txt
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
idna==2.8
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.0
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pyzmq==18.0.2
requests==2.22.0
scipy==1.3.0
six==1.12.0
torch==1.1.0.post2
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.3.0
tornado==6.0.3
urllib3==1.25.3
visdom==0.1.8.8
websocket-client==0.56.0
